refactor(mcp-repl): split library core from binary
